third_party/blink/web_tests/platform/generic/fast/frames/frame-js-url-clientWidth-expected.txt

ALERT: This page tests whether an iframe correctly reports renderer-dependent values when executing a javascript: load. If the test passes, you'll see a series of 'PASS' messages below.

ALERT: PASS: frame.frameElement.getBoundingClientRect().width should be 250 and is.

ALERT: PASS: frame.frameElement.clientWidth should be 250 and is.

ALERT: PASS: frame.frameElement.getBoundingClientRect().height should be 250 and is.

ALERT: PASS: frame.frameElement.clientHeight should be 250 and is.
